rename molecular to molecule
