+++ pyyabc/molecule/all_perms.txt
