+++ pyyabc/molecule/symmetry.py
