+++ pyyabc/molecule/tests/test_structure.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy
 
-from pyyabc.molecular.structure import Molecular
+from pyyabc.molecule.structure import Molecule
 
 
 class TestMol(unittest.TestCase):
@@ -9,7 +9,7 @@ class TestMol(unittest.TestCase):
     def test_init(self):
         positions = [0, 0, 0]
         atoms = ['NaN_10']
-        mole = Molecular(positions, atoms)
+        mole = Molecule(positions, atoms)
         self.assertEqual(mole.atoms.tolist(), [1010])
 
     def test_get_permutations(self):
@@ -19,7 +19,7 @@ class TestMol(unittest.TestCase):
                            0.3345, -0.9314, -0.4496,
                            -1.0685, -0.0537, 0.1921]).reshape((-1, 3))
         atoms = ['H', 'C', 'H', 'H', 'H']
-        methane = Molecular(pos, atoms)
+        methane = Molecule(pos, atoms)
 
         symprec = 1e-2
         perm = methane.get_symmetry_permutation(symprec)
@@ -34,7 +34,7 @@ class TestMol(unittest.TestCase):
                    -2.0, 0.0, 0.0,
                    0.0, 0.0, 0.0]
         cco_atoms = [6, 6, 8]
-        mol = Molecular(cco_pos, cco_atoms)
+        mol = Molecule(cco_pos, cco_atoms)
         self.assertTrue(mol.check(elements=None, limit=0.1))
 
         # 测试元素c-c距离过近, 但只测o元素时候不会过近
@@ -42,7 +42,7 @@ class TestMol(unittest.TestCase):
                    2.01, 0.0, 0.0,
                    0.0, 0.0, 0.0]
         cco_atoms = [6, 6, 8]
-        mol = Molecular(cco_pos, cco_atoms)
+        mol = Molecule(cco_pos, cco_atoms)
         self.assertFalse(mol.check(elements=['C'], limit=0.1))
         self.assertTrue(mol.check(elements=['O'], limit=0.1))
